06-database-integration-mongoDB/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
pydantic
python-dotenv